refactor and add domain setting ui in modal
