+++ app/Policies/WorkspacePolicy.php
@@ -88,10 +88,6 @@ class WorkspacePolicy
      */
     public function workspaceAdmin(User $user, Workspace $workspace)
     {
-        if(UserWorkspace::where('user_id', $user->id)->where('workspace_id', $workspace->id)->first()->role == 'admin') {
-            return true;
-        }else{
-            return false;
-        }
+        return UserWorkspace::where('user_id', $user->id)->where('workspace_id', $workspace->id)->first()->role == 'admin'; 
     }
 }
